Rename 'DispatcherActor' to 'Dispatcher'
